Update getting-started-with-ios-apps.md
When I followed the Getting Started with iOS apps, I was stuck for more than a day trying to figure out why a recognized simulator was failing. It was quite a bad first-impression.
It would have been a lot more comfortable if the documentation had hinted that I should check the Stack tab for the XCode version.

When I made a new iOS project to test Bitrise, it defaulted to iOS 12.4, because I had upgraded to Xcode 10.3, but this Stack had apparently defaulted to 10.2, which did not support 12.4 and was therefore failing with the missing simulator.