fix: correct wrong information (#496)
See https://cordova.apache.org/docs/en/latest/guide/platforms/ios/#signing-an-app

If the provisioning profile is set a manual provisioning is assumed, which makes the build fail.

The code signing identity should be "iPhone Developer" starting from Xcode 8. This version does not work on modern OS X anymore, which is why I just changed the config and did not leave a note.