Adding warning as to where environment variables should be put on api payload
I've actually been stuck trying to use the build trigger api to pass a variable into another workflow. For about 8 hours straight i was digging through the bitrise docs and forum discussions without noticing what i was doing wrong, when i stumbled upon this post:

https://devcenter.bitrise.io/en/api/triggering-and-aborting-builds.html

Which solved it. It's from 2020 and the guys mention that updating the docs would help a lot, which i agree, so here i am.